third_party/WebKit/LayoutTests/fast/html/imports/no-browsing-context.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This ensures that @import doesn't work with headless documents.")
window.jsTestIsAsync = true;
var xhr = new XMLHttpRequest();
xhr.open("GET", "resources/root.html");
xhr.responseType = 'document';
xhr.send(null);
xhr.onload = function() {
    shouldBeNull("xhr.response.getElementById('hello').import");
    finishJSTest();
};
</script>
</body>
</html>
